pyproject.toml
[project]
name = "analyse-2-0"
version = "0.1.0"
description = ""
authors = [
    {name = "skip"}
]
requires-python = ">=3.13, <4.0"
dependencies = [
    "langchain",
    "langchain-core",
    "pillow",
    "langchain-openai",
    "python-dotenv"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
